requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.31.2
anyio==4.4.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.15.0
channels==4.0.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==15.1.0
cryptography==44.0.0
cssselect2==0.7.0
daphne==4.1.2
distlib==0.3.9
distro==1.9.0
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
duckduckgo_search==6.2.1
eventlet==0.36.1
fastlogging==1.0.0
filelock==3.15.4
Flask==2.0.1
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
frozenlist==1.5.0
fsspec==2024.6.1
gevent==24.2.1
gevent-websocket==0.10.1
git-filter-repo==2.45.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
groq==0.9.0
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.0
hyperlink==21.0.0
idna==3.3
incremental==24.7.2
iniconfig==2.0.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
keybert==0.8.5
keyboard==0.13.5
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
markdownify==0.13.1
MarkupSafe==2.1.5
mdurl==0.1.2
mistletoe==1.4.0
mistralai==0.4.2
mkl==2021.4.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
netlify-py==0.1.0
networkx==3.3
numpy==1.26.4
ollama==0.3.0
openai==1.36.1
orjson==3.10.6
oscrypto==1.3.0
packaging==24.1
passgen==1.1.1
pdfminer.six==20240706
pika==1.3.2
pillow==10.4.0
pipenv==2024.2.0
platformdirs==4.3.6
playwright==1.45.0
pluggy==1.5.0
propcache==0.2.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.1.0
Pygments==2.18.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.10.0
pynput==1.7.7
pyOpenSSL==24.3.0
pyparsing==3.1.2
pypdf==4.3.0
pypng==0.20220715.0
pyreqwest_impersonate==0.5.1
pytest==8.3.1
pytest-base-url==2.1.0
pytest-playwright==0.5.1
python-bidi==0.5.0
python-decouple==3.8
python-dotenv==1.0.1
python-engineio==4.9.1
python-slugify==8.0.4
python-socketio==5.11.3
pytz==2024.2
PyYAML==6.0.1
qrcode==7.4.2
redis==5.2.0
regex==2024.5.15
reportlab==4.2.2
requests==2.32.3
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
service-identity==21.1.0
setuptools==71.1.0
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlmodel==0.0.21
sqlparse==0.4.2
svglib==1.5.1
sympy==1.13.1
tbb==2021.13.0
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
torch==2.3.1
tqdm==4.66.4
transformers==4.42.4
twilio==9.3.7
Twisted==24.11.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.12.2
tzdata==2021.5
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.2
uv==0.2.27
virtualenv==20.27.0
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==6.8.2
wsproto==1.2.0
xhtml2pdf==0.2.16
yarl==1.18.0
zope.event==5.0
zope.interface==5.4.0
